+++ MapboxNavigation/NavigationViewController.swift
@@ -676,12 +676,17 @@ extension NavigationViewController: NavigationServiceDelegate {
         
         if !isConnectedToCarPlay, // CarPlayManager shows rating on CarPlay if it's connected
             service.routeProgress.isFinalLeg && advancesToNextLeg && showsEndOfRouteFeedback {
-            self.mapViewController?.showEndOfRoute { _ in }
+            showEndOfRouteFeedback()
         }
         return advancesToNextLeg
 
     }
     
+    @objc public func showEndOfRouteFeedback(duration: TimeInterval = 1.0, completionHandler: ((Bool) -> Void)? = nil) {
+        guard let mapController = mapViewController else { return }
+        mapController.showEndOfRoute(duration: duration, completion: completionHandler)
+    }
+
     @objc public func navigationService(_ service: NavigationService, willBeginSimulating progress: RouteProgress, becauseOf reason: SimulationIntent) {
         switch service.simulationMode {
         case .always: