+++ MapboxNavigationTests/Support/FBSnapshotTestCase.swift
@@ -2,11 +2,11 @@ import Foundation
 import FBSnapshotTestCase
 
 @nonobjc extension FBSnapshotTestCase {
-    func verify(_ view: UIView, file: StaticString = #file, line: UInt = #line) {
-        FBSnapshotVerifyView(view, suffixes: ["_64"], tolerance: 0.05, file: file, line: line)
+    func verify(_ view: UIView, overallTolerance: CGFloat = 0.05) {
+        FBSnapshotVerifyView(view, suffixes: ["_64"], overallTolerance: overallTolerance)
     }
     
-    func verify(_ layer: CALayer, file: StaticString = #file, line: UInt = #line) {
-        FBSnapshotVerifyLayer(layer, suffixes: ["_64"], tolerance: 0.05, file: file, line: line)
+    func verify(_ layer: CALayer, overallTolerance: CGFloat = 0.05) {
+        FBSnapshotVerifyLayer(layer, suffixes: ["_64"], overallTolerance: overallTolerance)
     }
 }